Move Icon to atoms
