+++ src/ui/atoms/icon.svelte
@@ -1,10 +1,10 @@
 <script module>
   import { type Component } from "svelte";
-  import { default as IconArrow } from "./icons/icon-arrow.svelte";
-  import { default as IconCart } from "./icons/icon-cart.svelte";
-  import { default as IconError } from "./icons/icon-error.svelte";
-  import { default as IconLock } from "./icons/icon-lock.svelte";
-  import { default as IconSuccess } from "./icons/icon-success.svelte";
+  import { default as IconArrow } from "../atoms/icons/icon-arrow.svelte";
+  import { default as IconCart } from "../atoms/icons/icon-cart.svelte";
+  import { default as IconError } from "../atoms/icons/icon-error.svelte";
+  import { default as IconLock } from "../atoms/icons/icon-lock.svelte";
+  import { default as IconSuccess } from "../atoms/icons/icon-success.svelte";
 
   export type IconName = "arrow" | "cart" | "error" | "lock" | "success";
