+++ src/ui/atoms/icons/icon-cart.svelte
