+++ src/ui/atoms/icons/icon-error.svelte
