+++ src/ui/atoms/icons/icon-lock.svelte
