+++ src/ui/atoms/icons/icon-success.svelte
